Make the width and height of the chart configurable by props
Before this change, the tiles size was calculated based on the size of the container. The problem is that the component had to render the first time in order find out the size of the html element using javascript, then it was necessary to use setState in componentDidMount to rerender the component. To avoid the use of setState inside componentDidMount, we are now passing the width and height as property.