fix(overlay): Fixed corrupted state when call open() and close() to fast and often (#196)
The problem was that close removes the container height and the class mod-open once the animation is finished.
When open() is called within these 200ms the target height is set on the container and the class mod-open is added, then the animatiion finishes and the height and mod-open is removed again.
Afterwards close() can be called again but since there is no change (class already removed and height is 0) the animation end events are never fired and the listener is not removed. So this effect will be repeated on the next open() and close() call.